+++ src/main/java/nordmods/iobvariantloader/mixin/common/dragon/ADragonBaseFlyingRideableMixin.java
@@ -9,7 +9,7 @@ import net.minecraft.world.entity.ai.control.MoveControl;
 import net.minecraft.world.entity.ai.navigation.GroundPathNavigation;
 import net.minecraft.world.level.Level;
 import nordmods.iobvariantloader.IoBVariantLoader;
-import nordmods.iobvariantloader.util.AltLandNavigation;
+import nordmods.iobvariantloader.util.alt_navigation.AltLandNavigation;
 import nordmods.iobvariantloader.util.sound_redirect.SoundRedirectUtil;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;